refactor: rename plugin/ to plugins/ and add plugins, utils to package files
