+++ lib/react-typescript.js
@@ -1,4 +1,4 @@
-import reactPlugin from "../plugin/react/index.js";
+import reactPlugin from "../plugins/react/index.js";
 
 /**
  * @return { import("eslint").Linter.Config[] }
