+++ lib/react.js
@@ -1,6 +1,6 @@
 import jsxA11yPlugin from "eslint-plugin-jsx-a11y-x";
 import hooksPlugin from "eslint-plugin-react-hooks";
-import reactPlugin from "../plugin/react/index.js";
+import reactPlugin from "../plugins/react/index.js";
 
 /**
  * @return { import("eslint").Linter.Config[] }
