+++ plugins/react/index.js
