+++ plugins/react/rules/jsx-boolean-value.js
