+++ plugins/react/rules/jsx-curly-brace-presence.js
