+++ plugins/react/rules/jsx-no-duplicate-props.js
