+++ plugins/react/rules/jsx-pascal-case.js
