+++ plugins/react/rules/jsx-uses-vars.js
