+++ plugins/react/rules/no-multi-comp.js
