+++ plugins/react/rules/self-closing-comp.js
