+++ plugins/react/rules/void-dom-elements-no-children.js
