+++ test/plugin/react/rules/jsx-curly-brace-presence-test.js
@@ -1,5 +1,5 @@
 import { RuleTester } from "eslint";
-import rule from "../../../../plugin/react/rules/jsx-curly-brace-presence.js";
+import rule from "../../../../plugins/react/rules/jsx-curly-brace-presence.js";
 
 const ruleTester = new RuleTester({
   languageOptions: {
