+++ test/plugin/react/rules/jsx-filename-extension-test.js
@@ -1,5 +1,5 @@
 import { RuleTester } from "eslint";
-import rule from "../../../../plugin/react/rules/jsx-filename-extension.js";
+import rule from "../../../../plugins/react/rules/jsx-filename-extension.js";
 
 const ruleTester = new RuleTester({
   languageOptions: {
