+++ test/plugin/react/rules/jsx-max-props-per-line-test.js
@@ -1,5 +1,5 @@
 import { RuleTester } from "eslint";
-import rule from "../../../../plugin/react/rules/jsx-max-props-per-line.js";
+import rule from "../../../../plugins/react/rules/jsx-max-props-per-line.js";
 
 const ruleTester = new RuleTester({
   languageOptions: {
