+++ test/plugin/react/rules/no-unescaped-entities-test.js
@@ -1,5 +1,5 @@
 import { RuleTester } from "eslint";
-import rule from "../../../../plugin/react/rules/no-unescaped-entities.js";
+import rule from "../../../../plugins/react/rules/no-unescaped-entities.js";
 
 const ruleTester = new RuleTester({
   languageOptions: {
